Move ExampleGlobalProjectile to FallenSoul and rename it to WraithPurificationGlobalProjectile
